requirements.txt
# Minimal dependencies - core functionality works with standard library only!
# These are optional but recommended:

# Optional: For .env file support
python-dotenv>=1.0.0

# Optional: For accurate token counting
tiktoken>=0.5.0

# Optional: For OpenAI API integration (if you want to enhance the agent)
openai>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0

